Add files via upload
A python programme that from a set of provided DNA sequences returns the sequence with the highest GC percentage. This was my first time using biopython and unsurprisingly it is a lot easier to use for bioinformatics. Also gained experience of creating and using a specific computational biology environment using jupyter lab and conda. Big thanks to Camillo for his hints.