Add files via upload
In this piece of code after being given a file I am looking to find a motif within the DNA. This is the first piece of code I am truly proud of as I learnt how to take information from one file and use this information in my code. The struggles I found was making sure that overlapping motifs in the DNA sequence were found, for example "ATATATAT" not just returning that there are two motifs of "ATAT" but three due to the overlap.